requirements.txt
streamlit==1.39.0
tensorflow==2.17.0
pandas==2.2.3
numpy==1.26.4
pillow==10.4.0
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
